[backend] removing code style breaches
