+++ backend/src/forecastbox/domain/auth/__init__.py
@@ -1,0 +1,5 @@
+"""Manages the Auth domain -- which is not really a business domain,
+but a utility. But for hierarchy reason, it is placed here.
+
+Depends on no other domain, dependended on no other domain.
+Used from every route that affects user-owned data."""
