+++ backend/src/forecastbox/routes/admin.py
@@ -25,7 +25,7 @@ from pydantic import UUID4, BaseModel
 from sqlalchemy import delete, select, update
 
 from forecastbox.domain.admin import Release, get_local_release, get_most_recent_release, get_pylock, mark_release, save_pylock
-from forecastbox.entrypoint.auth.users import current_active_user
+from forecastbox.domain.auth.users import current_active_user
 from forecastbox.schemata.user import UserRead, UserTable, UserUpdate, async_session_maker
 from forecastbox.utility.config import BackendAPISettings, CascadeSettings, ProductSettings, config
 from forecastbox.utility.rsjf import ExportedSchemas, FormDefinition, from_pydantic
