+++ backend/src/forecastbox/routes/auth.py
@@ -16,8 +16,8 @@ Responsible for all user authentication.
 PREFIX = "/api/v1"
 from fastapi import APIRouter
 
-from forecastbox.entrypoint.auth.oidc import oauth_client
-from forecastbox.entrypoint.auth.users import auth_backend, fastapi_users
+from forecastbox.domain.auth.oidc import oauth_client
+from forecastbox.domain.auth.users import auth_backend, fastapi_users
 from forecastbox.schemata.user import UserCreate, UserRead, UserUpdate
 from forecastbox.utility.config import config
