+++ backend/src/forecastbox/routes/experiment.py
@@ -26,9 +26,9 @@ from fastapi.exceptions import HTTPException
 from pydantic import BaseModel, PositiveInt
 
 import forecastbox.domain.experiment.service as experiment_service
+from forecastbox.domain.auth.users import get_auth_context
 from forecastbox.domain.experiment.exceptions import ExperimentAccessDenied, ExperimentNotFound, ExperimentVersionConflict, SchedulerBusy
 from forecastbox.domain.experiment.scheduling.background import start_scheduler, stop_scheduler
-from forecastbox.entrypoint.auth.users import get_auth_context
 from forecastbox.schemata.jobs import ExperimentDefinition
 from forecastbox.utility.auth import AuthContext
 from forecastbox.utility.pagination import PaginationSpec
