+++ backend/src/forecastbox/routes/run.py
@@ -34,9 +34,9 @@ from pydantic import BaseModel
 
 import forecastbox.domain.run.db as run_db
 import forecastbox.domain.run.service as run_service
+from forecastbox.domain.auth.users import get_auth_context
 from forecastbox.domain.run.cascade import ProductToOutputId, encode_result
 from forecastbox.domain.run.exceptions import RunAccessDenied, RunNotFound
-from forecastbox.entrypoint.auth.users import get_auth_context
 from forecastbox.routes.gateway import Globals
 from forecastbox.utility.auth import AuthContext
 from forecastbox.utility.config import config
